Deploy to heroku
Deploying to heroku as a docker file due to issues with needing to
compile nanoc and run npm install not working with the static build
pack.

Nginx serves the output directory.

Had to do some odd stuff to pick up the dynamic port on heroku see the
CMD in the dockerfile (heroku sets the $PORT with the current port).